+++ packages/web/src/javascripts/App.tsx
@@ -26,7 +26,7 @@ declare global {
 }
 
 import { IsWebPlatform, WebAppVersion } from '@/Constants/Version'
-import { DesktopManagerInterface, SNLog } from '@standardnotes/snjs'
+import { DesktopManagerInterface, Environment, Platform, SNLog } from '@standardnotes/snjs'
 import ApplicationGroupView from './Components/ApplicationGroupView/ApplicationGroupView'
 import { WebDevice } from './Application/Device/WebDevice'
 import { StartApplication } from './Application/Device/StartApplication'
@@ -42,10 +42,28 @@ const getKey = () => {
 }
 
 const setViewportHeight = () => {
-  document.documentElement.style.setProperty(
-    '--viewport-height',
-    `${visualViewport ? visualViewport.height : window.innerHeight}px`,
-  )
+  const currentValue = parseInt(document.documentElement.style.getPropertyValue('--viewport-height'))
+  const newValue = visualViewport && visualViewport.height > 0 ? visualViewport.height : window.innerHeight
+
+  if (currentValue && !newValue) {
+    return
+  }
+
+  if (!currentValue && !newValue) {
+    document.documentElement.style.setProperty('--viewport-height', '100vh')
+    return
+  }
+
+  document.documentElement.style.setProperty('--viewport-height', `${newValue}px`)
+}
+
+const setDefaultMonospaceFont = (platform?: Platform) => {
+  if (platform === Platform.Android) {
+    document.documentElement.style.setProperty(
+      '--sn-stylekit-monospace-font',
+      '"Roboto Mono", "Droid Sans Mono", monospace',
+    )
+  }
 }
 
 const startApplication: StartApplication = async function startApplication(
@@ -60,6 +78,9 @@ const startApplication: StartApplication = async function startApplication(
   let root: Root
 
   const onDestroy = () => {
+    if (device.environment === Environment.Desktop) {
+      window.removeEventListener('resize', setViewportHeight)
+    }
     window.removeEventListener('orientationchange', setViewportHeight)
     const rootElement = document.getElementById(ElementIds.RootId) as HTMLElement
     root.unmount()
@@ -74,8 +95,14 @@ const startApplication: StartApplication = async function startApplication(
     root = createRoot(appendedRootNode)
 
     disableIosTextFieldZoom()
+
     setViewportHeight()
     window.addEventListener('orientationchange', setViewportHeight)
+    if (device.environment === Environment.Desktop) {
+      window.addEventListener('resize', setViewportHeight)
+    }
+
+    setDefaultMonospaceFont(device.platform)
 
     root.render(
       <ApplicationGroupView