+++ packages/components/src/Packages/Editors/org.standardnotes.token-vault/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5](https://github.com/standardnotes/app/compare/@standardnotes/authenticator@2.2.4...@standardnotes/authenticator@2.2.5) (2022-10-05)
+
+**Note:** Version bump only for package @standardnotes/authenticator
+
 ## [2.2.4](https://github.com/standardnotes/app/compare/@standardnotes/authenticator@2.2.3...@standardnotes/authenticator@2.2.4) (2022-08-22)
 
 ### Bug Fixes