+++ packages/styles/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/standardnotes/app/compare/@standardnotes/styles@1.4.2...@standardnotes/styles@1.5.0) (2022-10-05)
+
+### Features
+
+* add free dark mode ([#1748](https://github.com/standardnotes/app/issues/1748)) ([09b994f](https://github.com/standardnotes/app/commit/09b994f8f938ae0536f42742f7a221df536c4c4a))
+
 ## [1.4.2](https://github.com/standardnotes/app/compare/@standardnotes/styles@1.4.1...@standardnotes/styles@1.4.2) (2022-06-30)
 
 ### Bug Fixes